refactor app.js and add header.js and main.js files
