+++ dom-merge-conflict/tasks/buttons-and-counter/src/header.js
@@ -1,0 +1,8 @@
+export function header(body) {
+  const header = document.createElement("header");
+  header.innerHTML = `
+            <h1>Number Counter</h1>
+            <p>A simple counter. Press increment to increase the count by one.</p>
+        `;
+  body.appendChild(header);
+}
