+++ dom-merge-conflict/tasks/buttons-and-counter/src/main.js
@@ -1,0 +1,8 @@
+export function main(body) {
+  const main = document.createElement("main");
+  main.innerHTML = `
+            <p id="counter" data-testid="counter">0</p>
+            <button id="increment">Increment</button>
+        `;
+  body.appendChild(main);
+}
